switch to readAsDataURL
The existing codebase uses readAsBinaryString which seems to fail in latest Chrome.

Trying to load ultrasound JPEG Baseline 8Bit Color images fails in latest chrome. This is especially problematic when trying to process frames of a multiframe file.

In general the readAsBinaryString is the wrong API to use here as the data already arrives as a URL so readAsDataURL works much better and is much simpler.